refactor: usefetchforms -> useForms
