+++ frontend/beCompliant/src/hooks/useForms.ts
@@ -3,9 +3,9 @@ import { axiosFetch } from '../api/Fetch';
 import { apiConfig } from '../api/apiConfig';
 import { Form } from '../api/types';
 
-export function useFetchForm(formId?: string) {
-  const queryKeys = apiConfig.form.queryKey(formId!);
-  const url = apiConfig.form.url(formId!);
+export function useForms() {
+  const queryKeys = apiConfig.forms.queryKey();
+  const url = apiConfig.forms.url();
 
   return useQuery({
     queryKey: queryKeys,
@@ -13,13 +13,16 @@ export function useFetchForm(formId?: string) {
     refetchOnMount: false,
     notifyOnChangeProps: ['data', 'error'],
     queryFn: () =>
-      axiosFetch<Form>({ url: url }).then((response) => response.data),
+      axiosFetch<Form[]>({ url: url }).then((response) => response.data),
     select: formatFormData,
-    enabled: !!formId,
   });
 }
 
-function formatFormData(data: Form) {
+function formatFormData(data: Form[]) {
+  return data.map(formatFormatData);
+}
+
+function formatFormatData(data: Form) {
   return {
     ...data,
     records: data.records.map((record) => {
